# BUG: bug in left join on multi-index with sort=True or nulls