@behzadnouri 2015-01-10
i could not find one. i noticed the issue when i saw on current master [`_get_multiindex_indexer`](https://github.com/pydata/pandas/blob/22f148620b18bb373412e42d399a97a5958a6312/pandas/tools/merge.py#L528) does not use the passed-in `sort` argument, and calls  into `get_group_index` without overflow check.